+++ tests/mfem.cpp
@@ -49,7 +49,7 @@ using namespace std;
 using namespace Eigen;
 using namespace spaND;
 
-int64_t main(int64_t argc, char *argv[])
+int main(int argc, char *argv[])
 {
     // 1. Parse command-line options.
     const char *mesh_file = "../data/star.mesh";
